LPD-55828 Moving the hide Recycle Bin logic frrom AssetLibraryResourceImpl to GroupModelListener
